+++ contentcuration/contentcuration/static/js/edit_channel/preview/hbtemplates/add_content_dialog.handlebars
@@ -1,0 +1,36 @@
+<div id="uploader" class="container">
+	 <div class="row uploader-row">
+	 	<div class="action-button create_option" id="create_topic">CREATE TOPIC</div>
+	 	<div class="action-button create_option" id="upload_file">UPLOAD</div>
+	 	<div class="action-button create_option" id="import_node">IMPORT</div>
+	 	<!--
+	 	<div class="col-xs-4 create_option construction" id="create_exercise">
+	 		<h1><span class="glyphicon glyphicon-star" aria-hidden="true"></span></h1>
+	 		Create Exercise
+	 	</div>
+	 	-->
+	 </div>
+	  <!--<h5>You can also drag and drop from your computer (under construction)</h5>-->
+
+	 <h4>Upload one or more files here</h4>
+	 <div class="row headers">
+	 	<div class="col-xs-2">Type</div>
+	 	<div class="col-xs-10">Name</div>
+ 	 </div>
+	 <ul class="list-unstyled" id="upload_content_add_list">
+	 {{#each node_list}}
+	 	<li id="content_item_{{cid}}" class="row {{kind}}"></li>
+	 {{/each}}
+	 </ul>
+	 <!--
+	 <h5>Add from a channel I'm subscribed to (under construction)</h4>
+	 <select class="construction">
+	 	<option>Khan Academy</option>
+	 </select>
+	 -->
+	 <div class="edit_metadata pull-right">
+		<!--<p>Edit Details</p>-->
+		<a class="action-text">NEXT</a>
+	 </div>
+
+</div>
