+++ contentcuration/contentcuration/static/js/edit_channel/preview/hbtemplates/add_content_header.handlebars
@@ -1,0 +1,2 @@
+<span class="glyphicon glyphicon-plus-sign pull-left" aria-hidden="true"></span>
+Adding to <em>{{title}}</em>
