+++ contentcuration/contentcuration/static/js/edit_channel/preview/hbtemplates/content_list_item.handlebars
@@ -1,0 +1,7 @@
+<div class="col-xs-2 content_type">
+	{{to_upper_case topic.attributes.kind}}
+</div>
+<div class="col-xs-10 name">
+	<input type="text" class="content_name" value="{{topic.attributes.title}}" maxlength="200"/>
+	<span class="glyphicon glyphicon-remove remove_topic pull-right" aria-hidden="true"></span>
+</div>
