+++ contentcuration/contentcuration/static/js/edit_channel/preview/hbtemplates/edit_metadata_dialog.handlebars
@@ -1,0 +1,65 @@
+<div id="metadata" class="container">
+	<div id="hidden_node"></div>
+	<!-- Optional, not needed if editing single files -->
+	<div id="topic_tree_selector"></div>
+	{{#if multiple_selected}}
+		<div class="section pull-left">
+			<ul class="list-unstyled" id="uploaded_list">
+			{{#each node_list}}
+				<li id="item_{{cid}}" class="uploaded disable_on_error"><li>
+			{{else}}
+				<li><h5 id="none-in-metadata-form">&nbsp;You didn't add<br>any content</h5></li>
+			{{/each}}
+			</ul>
+			{{#if allow_add}}
+				<a class="action-text" id="add_more_button">BACK</a>
+			{{/if}}
+		</div>
+	{{/if}}
+	<nav>
+		<a class="btn btn-tab" id="metadata_details_btn">Details</a>
+		<a class="btn btn-tab" id="metadata_preview_btn">Preview</a>
+	</nav>
+	<div  class="metadata_panel">
+		<div id="metadata_edit_details" class="input_area container-fluid">
+			<div class="section pull-left">
+				<h4 id="original_filename_area">Original Filename <i id="original_filename">No file selected.</i></h4>
+				<div id="editmetadata_format_section">
+                </div>
+				<h4>Title<label class="required">*<span id="title_error"></span></label></h4>
+				<input type="text" id="input_title" class="disable-on-edit upload_input" placeholder="Enter a title" maxlength="200" required/>
+				<h4>Description <i id="description_counter" class="pull-right">{{word_limit}} chars left</i></h4>
+				<textarea id="input_description" class="disable-on-edit upload_input" placeholder="Enter a description of your content" maxlength={{word_limit}}/></textarea>
+				<h4>Tags (optional)</h4>
+				<input type="text" id="tag_box" class="tag_input" placeholder="Press 'Enter' to add tag"/>
+				<div class="container-fluid">
+					<div class="row" id="tag_area">
+					</div>
+				</div>
+			</div>
+			<!--
+			<div class="section construction pull-right">
+				<h4>Related Content</h4>
+				<div class="content-relation-area">
+				</div>
+				<br/>
+				<h4>Prerequisite Content</h4>
+				<div class="content-relation-area">
+				</div>
+			</div>
+			-->
+		</div>
+
+		<div id="metadata_preview" class="container-fluid">
+
+		</div>
+	</div>
+	<br/>
+	<div class="metadata_save_buttons">
+		<div class="pull-right">
+			<span id="validating_text">Validating...</span>
+			<a id="upload_save_button" class="disable_on_error editmetadata_save">SAVE</a>
+			<a class="btn btn-default" id="upload_save_finish_button" class="disable_on_error editmetadata_save">SAVE &amp; FINISH</a>
+		</div>
+	</div>
+</div>
