+++ contentcuration/contentcuration/static/js/edit_channel/preview/hbtemplates/edit_metadata_header.handlebars
@@ -1,0 +1,2 @@
+<span class="glyphicon glyphicon-pencil pull-left" aria-hidden="true"></span>
+Editing Content Details
