+++ contentcuration/contentcuration/static/js/edit_channel/preview/hbtemplates/exercise_create.handlebars
@@ -1,0 +1,7 @@
+<span class="btn btn-default close_exercise_create">
+	<span>Cancel</span>
+</span>
+<span class="btn btn-success pull-right submit_exercise">
+	<i class="glyphicon glyphicon-ok" aria-hidden="true"></i>
+	<span>Save & Add Exercise</span>
+</span>
