+++ contentcuration/contentcuration/static/js/edit_channel/preview/hbtemplates/import_dialog.handlebars
@@ -1,0 +1,12 @@
+<div id="import_from_channel_box"></div>
+<br/>
+<div id="import_bottom_container">
+	<a class="action-text" data-dismiss="modal">
+		<span>CANCEL</span>
+	</a>
+	<button disabled class="action-button pull-right" id="import_content_submit">
+		<i class="glyphicon glyphicon-ok" aria-hidden="true"></i>
+		<span> IMPORT SELECTED</span>
+	</button>
+	<span id="import_file_count" class="pull-right"><em>No files selected</em></span>
+</div>
