+++ contentcuration/contentcuration/static/js/edit_channel/preview/hbtemplates/import_list.handlebars
@@ -1,0 +1,3 @@
+<ul class="list-unstyled import-list {{#if is_channel}}first_list{{/if}}" id="import_list_{{index}}">
+	{{#if is_empty}}<em> No content found. </em>{{/if}}
+</ul>
