+++ contentcuration/contentcuration/static/js/edit_channel/preview/hbtemplates/import_modal.handlebars
@@ -1,0 +1,12 @@
+<div class="modal fade" id="importingmodalback">
+  <div class="modal-dialog">
+    <div class="modal-content import_modal_content">
+      <div class="modal-header">
+      <button type="button" class="close" data-dismiss="modal" aria-label="Close"><span aria-hidden="true">&times;</span></button>
+        <h4 class="modal-title">Import from Another Channel</h4>
+      </div>
+      <div class="modal-body">
+      </div>
+    </div>
+  </div>
+</div>
