+++ contentcuration/contentcuration/static/js/edit_channel/preview/hbtemplates/preview_dialog.handlebars
@@ -1,0 +1,5 @@
+<div id="preview_window">
+	<div class="default_preview container-fluid">
+		<h5 class="default_preview_text text-center"><em>No preview to display.</em></h5>
+	</div>
+</div>
