+++ contentcuration/contentcuration/static/js/edit_channel/preview/hbtemplates/preview_templates/audio.handlebars
@@ -1,0 +1,4 @@
+<audio controls class="audio_preview">
+  <source src="{{source}}" type="audio/{{extension}}">
+	Your browser does not support the audio element.
+</audio>
