+++ contentcuration/contentcuration/static/js/edit_channel/preview/hbtemplates/preview_templates/default.handlebars
@@ -1,0 +1,3 @@
+<div class="default_preview container-fluid">
+	<h5 class="default_preview_text text-center"><em>No preview for {{title}}</em></h5>
+</div>
