+++ contentcuration/contentcuration/static/js/edit_channel/preview/hbtemplates/preview_templates/document.handlebars
@@ -1,0 +1,1 @@
+<embed src="{{source}}" class="pdf_preview">
