+++ contentcuration/contentcuration/static/js/edit_channel/preview/hbtemplates/preview_templates/image.handlebars
@@ -1,0 +1,1 @@
+<img src="{{source}}" alt="Error loading image" class="image_preview"/>
