+++ contentcuration/contentcuration/static/js/edit_channel/preview/hbtemplates/preview_templates/video.handlebars
@@ -1,0 +1,4 @@
+<video class="video_preview" controls>
+  	<source src="{{source}}" type="video/{{extension}}">
+	Your browser does not support the video tag.
+</video>
