+++ contentcuration/contentcuration/static/js/edit_channel/preview/hbtemplates/uploaded_list_item.handlebars
@@ -1,0 +1,9 @@
+<input type="checkbox" id="check_{{id}}" class="upload_item_checkbox"/>
+<label for="check_{{id}}" class="uploaded_list_item">
+	<!--<span class="glyphicon glyphicon-remove remove_item pull-right" aria-hidden="true"></span>-->
+	<h6>{{to_upper_case topic.attributes.kind}}</h6>
+	<h5 class="truncate">
+		<span class="item_name" id="{{id}}">{{topic.attributes.title}}</span>
+		<!--<span class="glyphicon glyphicon-chevron-right" aria-hidden="true"></span>-->
+	</h5>
+</label>
