+++ contentcuration/contentcuration/static/js/edit_channel/preview/hbtemplates/uploader_modal.handlebars
@@ -1,0 +1,12 @@
+<div class="modal fade" id="uploaderModal">
+  <div class="modal-dialog">
+    <div class="modal-content upload-modal-content">
+      <div class="modal-header">
+        <button type="button" class="close" data-dismiss="modal" aria-label="Close"><span aria-hidden="true">&times;</span></button>
+        <h4 class="modal-title truncate"></h4>
+      </div>
+      <div class="modal-body">
+      </div>
+    </div>
+  </div>
+</div>
